Move main() entry to separate project
